roles list removed; users now only have an activerole field. redirects users to enter password page when not logged in

+++ RepairCafe/RepairCafe/views.py
@@ -63,7 +63,7 @@ def send_queue_update(group_name, queue_name, update_type):
         },
     )
 
-@login_required
+
 def index(request):
     print(request.user.activerole)
     return render(request, 'RepairCafe/index.html', context={})
@@ -82,8 +82,12 @@ def get_queue_position(request, repairNumber):
     except Ticket.DoesNotExist:
         return JsonResponse({'error': 'Ticket not found'}, status=404)
 
-@login_required
-def reset_data(request):
+
+def reset_data(request): 
+
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+    
     script.populate()
     referer_url = request.META.get('HTTP_REFERER')
     if referer_url:
@@ -94,16 +98,15 @@ def reset_data(request):
 
 def main_queue(request):
 
-    print(request.user.roles)
-    print(request.user.username)
+    print("Session Data:", request.session.items())
 
-    if len(request.user.roles) > 1 or not request.user.activerole: #if user has multiple roles, redirect to a page where they select the role they use to access this page
-        redirect('RepairCafe:role_selection')
-    if len(request.user.roles) == 1: #defaults to the only role available in case a user logs out of their active role
-        request.user.activerole = request.user.roles[0]
-        request.user.save()
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+    
+    #MUST INCLUDE REDIRECTION TO PREVIOUS PAGE FOR VISITORS
     
     
+    print(request.user.activerole)
 
 
 
@@ -140,17 +143,20 @@ def main_queue(request):
 
 def waiting_list(request):
 
-    print(request.user.roles)
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+
+    
     print(request.user.activerole)
 
     if not request.user.is_authenticated:
        return redirect('RepairCafe:enter_password')
     
-    if len(request.user.roles) > 1 and not request.user.activerole: #if user has multiple roles and activerole is not set, redirect to a page where they select the role they use to access this page
+    '''if len(request.user.roles) > 1 and not request.user.activerole: #if user has multiple roles and activerole is not set, redirect to a page where they select the role they use to access this page
        return redirect('RepairCafe:role_selection')
     elif len(request.user.roles) == 1:
         request.user.activerole = request.user.roles[0]
-        request.user.save()
+        request.user.save()'''
 
 
 
@@ -181,13 +187,14 @@ def waiting_list(request):
 
 def checkout_queue(request, activerole=""):
 
-    if len(request.user.roles) > 1 and not request.user.activerole: #if user has multiple roles, redirect to a page where they select the role they use to access this page
+    '''if len(request.user.roles) > 1 and not request.user.activerole: #if user has multiple roles, redirect to a page where they select the role they use to access this page
         redirect('RepairCafe:role_selection')
     elif len(request.user.roles) == 1:
         request.user.activerole = request.user.roles[0]
-        request.user.save()
+        request.user.save()'''
 
-
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
 
     context_dict={}
     try:
@@ -212,6 +219,9 @@ def checkout_queue(request, activerole=""):
 
 def basic_stats(request):
 
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+
     checkedin = Ticket.objects.exclude(repairStatus="WAITING").count()
     checkedout = Ticket.objects.filter(isCheckedOut=True).count()
     successful = Ticket.objects.filter(repairStatus="COMPLETED").count()
@@ -230,6 +240,10 @@ def basic_stats(request):
 
 
 def accept_ticket(request, repairNumber):
+
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+    
     ticket = Ticket.objects.get(repairNumber=repairNumber)
     repairNumber = repairNumber
 
@@ -247,6 +261,10 @@ def accept_ticket(request, repairNumber):
 
 
 def repair_ticket(request, repairNumber):
+
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+    
     ticket = get_object_or_404(Ticket, repairNumber=repairNumber)
 
     repairer_name = request.session.get('repairer_name', None)
@@ -268,6 +286,10 @@ def repair_ticket(request, repairNumber):
 
 
 def mark_incomplete_ticket(request, repairNumber):
+
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+    
     ticket = get_object_or_404(Ticket, repairNumber=repairNumber)
     if request.method == 'POST':
         incompleteForm = IncompleteTicketForm(request.POST)
@@ -291,6 +313,10 @@ def mark_incomplete_ticket(request, repairNumber):
 
 
 def repair_item(request, repairNumber):
+
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+    
     context_dict = {}
     ticket = Ticket.objects.get(repairNumber=repairNumber)
     incompleteForm = IncompleteTicketForm()
@@ -304,6 +330,10 @@ def repair_item(request, repairNumber):
 
 
 def complete_ticket(request, repairNumber):
+
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+    
     ticket = Ticket.objects.get(repairNumber=repairNumber)
     if ticket.repairStatus == 'BEING_REPAIRED' and ticket.itemCategory == "ELECM":
         ticket.complete_ticket()
@@ -326,6 +356,10 @@ def complete_ticket(request, repairNumber):
 
 
 def pat_test(request, repairNumber):
+
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+    
     if request.method == 'POST':
         ticket = get_object_or_404(Ticket, repairNumber=repairNumber)
         action = request.POST.get('action')
@@ -358,6 +392,10 @@ def pat_test(request, repairNumber):
 
 
 def delete_ticket(request,repairNumber):
+
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+    
     ticket = Ticket.objects.get(repairNumber=repairNumber)
     ticket.delete_ticket()
     messages.success(request, f"Ticket: {ticket.itemName}, has been removed")
@@ -366,6 +404,10 @@ def delete_ticket(request,repairNumber):
 
 
 def checkout_ticket(request, repairNumber):
+
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+    
     ticket = get_object_or_404(Ticket, repairNumber=repairNumber)
     if ticket.repairStatus == 'COMPLETED' or ticket.repairStatus == 'INCOMPLETE':
         ticket.checkout()
@@ -382,6 +424,10 @@ def checkout_ticket(request, repairNumber):
 
 
 def change_category(request, repairNumber):
+
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+    
     ticket = get_object_or_404(Ticket, repairNumber=repairNumber)
     if request.method == 'POST':
         new_category = request.POST.get('new_category')
@@ -398,13 +444,14 @@ def change_category(request, repairNumber):
 
 def authenticate_roles(request):
 
-    new_roles = UserRoles.objects.create(roles=[])
+    new_roles = UserRoles.objects.create()
     login(request, new_roles,backend='django.contrib.auth.backends.ModelBackend')
     
     return new_roles
 
 def enter_password(request):
       
+    
     if not request.user.is_authenticated:
      user = authenticate_roles(request)  
         
@@ -425,28 +472,20 @@ def enter_password(request):
             role = "volunteer"
             request.session['preset_password_verified'] = True
         else:
-            return render(request, 'RepairCafe/enter_password.html', {'error': 'Incorrect Password'})
-
-    #very messy here but basically, visitors can only log in once, volunteers and repairers can log in a second time to get the other of those two roles
-    #can be cleaned up later
-        if role == "visitor" and request.user.roles == ["visitor"]:
-         return render(request, 'RepairCafe/enter_password.html', {'error': 'Visitors can only log in once at a time.'})
-        
-        else:
+            return render(request, 'RepairCafe/enter_password.html', {'error': 'Incorrect Password'})        
          
-         if role not in request.user.roles:
-            print("assigning role")
-            request.user.roles.append(role)
-            request.user.activerole = role
-            request.user.save()
+        
+        print("assigning role")
+        request.user.activerole = role
+        request.user.save()
 
         
-        if role == "visitor" and len(request.user.roles) == 1:
+        if role == "visitor":
             return redirect('RepairCafe:house_rules')
         elif role == "volunteer":
             return redirect('RepairCafe:waiting_list')
         elif role == "repairer":
-            return redirect('RepairCafe:main_queue')
+            return redirect('RepairCafe:repairer_login')
         
 
         
@@ -454,10 +493,18 @@ def enter_password(request):
     return render(request, 'RepairCafe/enter_password.html')
 
 def repairer_login(request):
+
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+    
     context_dict = {}
     repairers = Repairer.objects.all()
     if request.method == "POST":
         selected_repairer_name = request.POST.get('selected_repairer')
+        user = authenticate_roles(request)
+        request.user.activerole = "repairer"
+        #request.user.roles.append("repairer")
+        request.user.save()
         if not selected_repairer_name:
             context_dict['errors'] = "Error: Please select a repairer before confirming."
         else:
@@ -481,6 +528,10 @@ def settings_page(request):
     return render(request, 'RepairCafe/settings_page.html')
 
 def volunteer_checkout(request, repairNumber):
+
+    if (not request.user.is_authenticated) or not request.user.activerole:
+        return redirect(reverse('RepairCafe:enter_password'))
+    
     ticket = get_object_or_404(Ticket, repairNumber=repairNumber)
     if ticket.repairStatus != "COMPLETED" and ticket.repairStatus != "INCOMPLETE":
         raise Http404("The ticket is not in the desired state.")
@@ -507,7 +558,7 @@ def volunteer_checkout_success(request):
 
 
 
-def role_selection(request): #users with multiple roles are redirected here and made to choose a role
+def role_selection(request): #OBSOLETE
 
 
     if request.method == 'POST':
@@ -532,13 +583,17 @@ def logout(request):
         selected = request.POST.get('role')
         if selected:
             print(selected)      
-            request.user.roles.remove(selected)     
+            #request.user.roles.remove(selected)     
             request.user.activerole = ""
-            request.user.save()
+            request.user.delete()
+
+            '''if len(request.user.roles) == 0:
+                request.user.delete()
+                return redirect('RepairCafe:enter_password')'''
             
-            return redirect('RepairCafe:index')
+            return redirect('RepairCafe:main_queue')
         
-    return render(request, 'RepairCafe/logout.html', {'roles': request.user.roles})
+    return render(request, 'RepairCafe/logout.html', {'roles': request.user.activerole})
 
 """
 Visitor Flow
